Made Every Struct a different file
