+++ Listenbrainz/model/Track.swift
@@ -1,0 +1,15 @@
+//
+//  Track.swift
+//  Listenbrainz
+//
+//  Created by Gaurav Bhardwaj on 06/06/23.
+//
+
+import Foundation
+
+
+// MARK: - Track
+struct Track: Codable {
+    let name: String
+    let duration: Int
+}
